Revamp documentation parsing
