+++ compiler/ballerina-parser/src/main/java/io/ballerinalang/compiler/internal/parser/BallerinaParser.java
@@ -672,7 +672,7 @@ public class BallerinaParser extends AbstractParser {
         switch (tokenKind) {
             case EOF_TOKEN:
                 return null;
-            case HASH_TOKEN:
+            case DOCUMENTATION_CONTENT_STRING:
             case AT_TOKEN:
                 metadata = parseMetaData(tokenKind);
                 return parseTopLevelNode(metadata);
@@ -3246,7 +3246,7 @@ public class BallerinaParser extends AbstractParser {
                 STNode semicolonToken = parseSemicolon();
                 endContext();
                 return STNodeFactory.createTypeReferenceNode(asterisk, type, semicolonToken);
-            case HASH_TOKEN:
+            case DOCUMENTATION_CONTENT_STRING:
             case AT_TOKEN:
                 startContext(ParserRuleContext.RECORD_FIELD);
                 STNode metadata = parseMetaData(nextTokenKind);
@@ -4850,7 +4850,7 @@ public class BallerinaParser extends AbstractParser {
             case CONST_KEYWORD:
             case LISTENER_KEYWORD:
             case EQUAL_TOKEN:
-            case HASH_TOKEN:
+            case DOCUMENTATION_CONTENT_STRING:
             case AT_TOKEN:
             case AS_KEYWORD:
             case IN_KEYWORD:
@@ -5284,7 +5284,7 @@ public class BallerinaParser extends AbstractParser {
             case FUNCTION_KEYWORD:
                 metadata = createEmptyMetadata();
                 break;
-            case HASH_TOKEN:
+            case DOCUMENTATION_CONTENT_STRING:
             case AT_TOKEN:
                 metadata = parseMetaData(nextTokenKind);
                 nextTokenKind = peek().kind;
@@ -5979,7 +5979,7 @@ public class BallerinaParser extends AbstractParser {
             case READONLY_KEYWORD:
                 return false;
             case EOF_TOKEN:
-            case HASH_TOKEN:
+            case DOCUMENTATION_CONTENT_STRING:
             case AT_TOKEN:
             case CLOSE_BRACE_TOKEN:
             case SEMICOLON_TOKEN:
@@ -6538,7 +6538,7 @@ public class BallerinaParser extends AbstractParser {
             case FUNCTION_KEYWORD:
                 metadata = createEmptyMetadata();
                 break;
-            case HASH_TOKEN:
+            case DOCUMENTATION_CONTENT_STRING:
             case AT_TOKEN:
                 metadata = parseMetaData(nextTokenKind);
                 nextTokenKind = peek().kind;
@@ -7099,8 +7099,8 @@ public class BallerinaParser extends AbstractParser {
         STNode docString;
         STNode annotations;
         switch (nextTokenKind) {
-            case HASH_TOKEN:
-                docString = parseDocumentationString();
+            case DOCUMENTATION_CONTENT_STRING:
+                docString = parseDocumentationContentString();
                 annotations = parseAnnotations();
                 break;
             case AT_TOKEN:
@@ -10126,7 +10126,7 @@ public class BallerinaParser extends AbstractParser {
             case EOF_TOKEN:
             case RESOURCE_KEYWORD:
             case LISTENER_KEYWORD:
-            case HASH_TOKEN:
+            case DOCUMENTATION_CONTENT_STRING:
             case PRIVATE_KEYWORD:
             case RETURNS_KEYWORD:
             case SERVICE_KEYWORD:
@@ -11415,7 +11415,7 @@ public class BallerinaParser extends AbstractParser {
         STToken nextToken = peek();
         STNode metadata;
         switch (nextToken.kind) {
-            case HASH_TOKEN:
+            case DOCUMENTATION_CONTENT_STRING:
             case AT_TOKEN:
                 metadata = parseMetaData(nextToken.kind);
                 break;
@@ -12994,260 +12994,12 @@ public class BallerinaParser extends AbstractParser {
         return errorCode;
     }
 
-    // --------------------------------- Documentation ---------------------------------
-
-    /*
-     * This section parses documentation.
-     * Ballerina flavored markdown (BFM) is supported by the documentation.
-     *
-     * During the parsing, parser will never go to the error handler.
-     * In case of an error, simply missing token will be returned.
-     */
-
-    /**
-     * Parse documentation string.
-     * <p>
-     * <code>
-     * DocumentationString :=
-     *          ( DocumentationLine
-     *          | ReferenceDocumentationLine
-     *          | DeprecationDocumentationLine
-     *          | ParameterDocumentationLine
-     *          | ReturnParameterDocumentationLine
-     *          | InvalidDocumentationLine ) +
-     * </code>
-     * <p>
-     * Refer {@link BallerinaLexer#readDocumentationToken}
-     *
-     * @return Parsed node
-     */
-    private STNode parseDocumentationString() {
-        List<STNode> docLines = new ArrayList<>();
-        STToken nextToken = peek();
-        while (nextToken.kind == SyntaxKind.HASH_TOKEN) {
-            docLines.add(parseSingleDocumentationLine());
-            nextToken = peek();
-        }
-
-        STNode documentationLines = STNodeFactory.createNodeList(docLines);
-        return STNodeFactory.createDocumentationStringNode(documentationLines);
-    }
-
-    /**
-     * Parse a single documentation line.
-     *
-     * @return Parsed node
-     */
-    private STNode parseSingleDocumentationLine() {
-        STNode hashToken = consume();
-        STToken nextToken = peek();
-        if (nextToken.kind == SyntaxKind.DOC_PLUS_TOKEN) {
-            return parseParameterDocumentationLine(hashToken);
-        } else if (nextToken.kind == SyntaxKind.DEPRECATION_LITERAL) {
-            return parseDeprecationDocumentationLine(hashToken);
-        }
-        return parseDocumentationLine(hashToken);
-    }
-
-    /**
-     * Parse deprecation documentation line.
-     *
-     * @param hashToken Hash token at the beginning of the line
-     * @return Parsed node
-     */
-    private STNode parseDeprecationDocumentationLine(STNode hashToken) {
-        STNode deprecationLiteral = consume();
-
-        List<STNode> docElements = parseDocumentationElements();
-        docElements.add(0, deprecationLiteral);
-
-        STNode docElementList = STNodeFactory.createNodeList(docElements);
-        return createDeprecationDocumentationLineNode(hashToken, docElementList);
-    }
-
-    /**
-     * Parse documentation line and reference documentation line.
-     *
-     * @param hashToken Hash token at the beginning of the line
-     * @return Parsed node
-     */
-    private STNode parseDocumentationLine(STNode hashToken) {
-        List<STNode> docElements = parseDocumentationElements();
-        STNode docElementList = STNodeFactory.createNodeList(docElements);
-
-        switch (docElements.size()) {
-            case 0:
-                // When documentation line is only a `#` token
-                return createDocumentationLineNode(hashToken, docElementList);
-            case 1:
-                STNode docElement = docElements.get(0);
-                if (docElement.kind == SyntaxKind.DOCUMENTATION_DESCRIPTION) {
-                    return createDocumentationLineNode(hashToken, docElementList);
-                }
-                // Else fall through
-            default:
-                return createReferenceDocumentationLineNode(hashToken, docElementList);
-        }
-    }
-
-    private List<STNode> parseDocumentationElements() {
-        List<STNode> docElements = new ArrayList<>();
-        STNode docElement;
-        SyntaxKind nextTokenKind = peek().kind;
-        while (!isEndOfIntermediateDocumentation(nextTokenKind)) {
-            if (nextTokenKind == SyntaxKind.DOCUMENTATION_DESCRIPTION) {
-                docElement = consume();
-
-            } else {
-                docElement = parseDocumentationReference();
-            }
-            docElements.add(docElement);
-            nextTokenKind = peek().kind;
-        }
-        return docElements;
-    }
-
-    private STNode parseDocumentationReference() {
-        STNode referenceType = STNodeFactory.createEmptyNode();
-        if (isDocumentReferenceType(peek().kind)) {
-            referenceType = consume();
-        }
-
-        STNode startBacktick = parseDocumentationBacktickToken();
-        STNode backtickContent = parseDocumentationBacktickContent();
-        STNode endBacktick = parseDocumentationBacktickToken();
-
-        return STNodeFactory.createDocumentationReferenceNode(referenceType, startBacktick, backtickContent,
-                endBacktick);
-    }
-
-    private boolean isDocumentReferenceType(SyntaxKind kind) {
-        switch (kind) {
-            case TYPE_DOC_REFERENCE_TOKEN:
-            case SERVICE_DOC_REFERENCE_TOKEN:
-            case VARIABLE_DOC_REFERENCE_TOKEN:
-            case VAR_DOC_REFERENCE_TOKEN:
-            case ANNOTATION_DOC_REFERENCE_TOKEN:
-            case MODULE_DOC_REFERENCE_TOKEN:
-            case FUNCTION_DOC_REFERENCE_TOKEN:
-            case PARAMETER_DOC_REFERENCE_TOKEN:
-            case CONST_DOC_REFERENCE_TOKEN:
-                return true;
-            default:
-                return false;
-        }
-    }
-
-    /**
-     * Parse parameter documentation line and return parameter documentation line.
-     *
-     * @param hashToken Hash token at the beginning of the line
-     * @return Parsed node
-     */
-    private STNode parseParameterDocumentationLine(STNode hashToken) {
-        STNode plusToken = consume();
-        STNode parameterName = parseDocumentationParameterName();
-        STNode dashToken = parseDocumentationDashToken();
-
-        List<STNode> docElements = parseDocumentationElements();
-        STNode docElementList = STNodeFactory.createNodeList(docElements);
-
-        SyntaxKind kind;
-        if (parameterName.kind == SyntaxKind.DOC_RETURN_KEYWORD) {
-            kind = SyntaxKind.RETURN_PARAMETER_DOCUMENTATION_LINE;
-        } else {
-            kind = SyntaxKind.PARAMETER_DOCUMENTATION_LINE;
-        }
-
-        return STNodeFactory.createParameterDocumentationLineNode(kind, hashToken, plusToken, parameterName, dashToken,
-                docElementList);
-    }
-
-    private boolean isEndOfIntermediateDocumentation(SyntaxKind kind) {
-        switch (kind) {
-            case DOCUMENTATION_DESCRIPTION:
-            case DOC_PLUS_TOKEN:
-            case DOC_PARAMETER_NAME:
-            case DOC_DASH_TOKEN:
-            case DOC_BACKTICK_TOKEN:
-            case DOC_BACKTICK_CONTENT:
-            case DOC_RETURN_KEYWORD:
-            case DEPRECATION_LITERAL:
-                return false;
-            default:
-                return !isDocumentReferenceType(kind);
-        }
-    }
-
-    /**
-     * Parse parameter name token.
-     *
-     * @return Parsed node
-     */
-    private STNode parseDocumentationParameterName() {
-        SyntaxKind tokenKind = peek().kind;
-        if (tokenKind == SyntaxKind.DOC_PARAMETER_NAME || tokenKind == SyntaxKind.DOC_RETURN_KEYWORD) {
-            return consume();
-        } else {
-            return STNodeFactory.createMissingToken(SyntaxKind.DOC_PARAMETER_NAME);
-        }
-    }
-
-    /**
-     * Parse documentation dash token.
-     *
-     * @return Parsed node
-     */
-    private STNode parseDocumentationDashToken() {
-        STToken token = peek();
-        if (token.kind == SyntaxKind.DOC_DASH_TOKEN) {
-            return consume();
-        } else {
-            return STNodeFactory.createMissingToken(SyntaxKind.DOC_DASH_TOKEN);
-        }
-    }
-
-    /**
-     * Parse documentation back-tick token.
-     *
-     * @return Parsed node
-     */
-    private STNode parseDocumentationBacktickToken() {
-        STToken token = peek();
-        if (token.kind == SyntaxKind.DOC_BACKTICK_TOKEN) {
-            return consume();
-        } else {
-            return STNodeFactory.createMissingToken(SyntaxKind.DOC_BACKTICK_TOKEN);
-        }
-    }
-
-    /**
-     * Parse documentation back-tick content token.
-     *
-     * @return Parsed node
-     */
-    private STNode parseDocumentationBacktickContent() {
-        STToken token = peek();
-        if (token.kind == SyntaxKind.DOC_BACKTICK_CONTENT) {
-            return consume();
-        } else {
-            return STNodeFactory.createMissingToken(SyntaxKind.DOC_BACKTICK_CONTENT);
-        }
-    }
-
-    private STNode createDocumentationLineNode(STNode hashToken, STNode documentationElements) {
-        return STNodeFactory.createDocumentationLineNode(SyntaxKind.DOCUMENTATION_LINE, hashToken,
-                documentationElements);
-    }
-
-    private STNode createDeprecationDocumentationLineNode(STNode hashToken, STNode documentationElements) {
-        return STNodeFactory.createDocumentationLineNode(SyntaxKind.DEPRECATION_DOCUMENTATION_LINE, hashToken,
-                documentationElements);
-    }
-
-    private STNode createReferenceDocumentationLineNode(STNode hashToken, STNode documentationElements) {
-        return STNodeFactory.createDocumentationLineNode(SyntaxKind.REFERENCE_DOCUMENTATION_LINE, hashToken,
-                documentationElements);
+    private STNode parseDocumentationContentString() {
+        STToken nextToken = consume();
+        TextDocument textDocument = TextDocuments.from(nextToken.toString());
+        AbstractTokenReader tokenReader = new TokenReader(new DocumentationLexer(textDocument.getCharacterReader()));
+        DocumentationParser documentationParser = new DocumentationParser(tokenReader);
+        return documentationParser.parse();
     }
 
     // ------------------------ Ambiguity resolution at statement start ---------------------------
