+++ compiler/ballerina-parser/src/main/java/io/ballerinalang/compiler/syntax/tree/SyntaxKind.java
@@ -196,12 +196,6 @@ public enum SyntaxKind {
     PARAMETER_DOC_REFERENCE_TOKEN(907, "parameter"),
     CONST_DOC_REFERENCE_TOKEN(908, "const"),
 
-    // Documentation syntax
-    DOC_PLUS_TOKEN(950, "+"),
-    DOC_DASH_TOKEN(951, "-"),
-    DOC_BACKTICK_TOKEN(952, "`"),
-    DOC_RETURN_KEYWORD(953, "return"),
-
     IDENTIFIER_TOKEN(1000),
     STRING_LITERAL(1001),
     DECIMAL_INTEGER_LITERAL(1002),
@@ -454,9 +448,10 @@ public enum SyntaxKind {
     DEPRECATION_DOCUMENTATION_LINE(4504),
     DOCUMENTATION_DESCRIPTION(4505),
     DOCUMENTATION_REFERENCE(4506),
-    DOC_PARAMETER_NAME(4507),
-    DOC_BACKTICK_CONTENT(4508),
+    PARAMETER_NAME(4507),
+    BACKTICK_CONTENT(4508),
     DEPRECATION_LITERAL(4509),
+    DOCUMENTATION_CONTENT_STRING(4510),
 
     INVALID(4),
     MODULE_PART(3),
