+++ compiler/ballerina-parser/src/test/java/io/ballerinalang/compiler/parser/test/ParserTestUtils.java
@@ -365,14 +365,15 @@ public class ParserTestUtils {
             case HEX_INTEGER_LITERAL:
             case DECIMAL_FLOATING_POINT_LITERAL:
             case HEX_FLOATING_POINT_LITERAL:
-            case DOC_PARAMETER_NAME:
-            case DOC_BACKTICK_CONTENT:
+            case PARAMETER_NAME:
+            case BACKTICK_CONTENT:
             case DEPRECATION_LITERAL:
                 return token.text();
             case XML_TEXT:
             case XML_TEXT_CONTENT:
             case TEMPLATE_STRING:
             case DOCUMENTATION_DESCRIPTION:
+            case DOCUMENTATION_CONTENT_STRING:
                 return cleanupText(token.text());
             default:
                 return token.kind.toString();
@@ -646,16 +647,6 @@ public class ParserTestUtils {
                 return SyntaxKind.PARAMETER_DOC_REFERENCE_TOKEN;
             case "CONST_DOC_REFERENCE_TOKEN":
                 return SyntaxKind.CONST_DOC_REFERENCE_TOKEN;
-
-            // Documentation syntax
-            case "DOC_PLUS_TOKEN":
-                return SyntaxKind.DOC_PLUS_TOKEN;
-            case "DOC_DASH_TOKEN":
-                return SyntaxKind.DOC_DASH_TOKEN;
-            case "DOC_BACKTICK_TOKEN":
-                return SyntaxKind.DOC_BACKTICK_TOKEN;
-            case "DOC_RETURN_KEYWORD":
-                return SyntaxKind.DOC_RETURN_KEYWORD;
 
             // Operators
             case "PLUS_TOKEN":
@@ -1208,12 +1199,14 @@ public class ParserTestUtils {
                 return SyntaxKind.DOCUMENTATION_DESCRIPTION;
             case "DOCUMENTATION_REFERENCE":
                 return SyntaxKind.DOCUMENTATION_REFERENCE;
-            case "DOC_PARAMETER_NAME":
-                return SyntaxKind.DOC_PARAMETER_NAME;
-            case "DOC_BACKTICK_CONTENT":
-                return SyntaxKind.DOC_BACKTICK_CONTENT;
+            case "PARAMETER_NAME":
+                return SyntaxKind.PARAMETER_NAME;
+            case "BACKTICK_CONTENT":
+                return SyntaxKind.BACKTICK_CONTENT;
             case "DEPRECATION_LITERAL":
                 return SyntaxKind.DEPRECATION_LITERAL;
+            case "DOCUMENTATION_CONTENT_STRING":
+                return SyntaxKind.DOCUMENTATION_CONTENT_STRING;
 
             // Trivia
             case "EOF_TOKEN":
